wpt_babel/content-security-policy/script-src/script-src-strict_dynamic_worker-importScripts.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js">
</script>
<script src="/resources/testharnessreport.js">
</script>
<script src="../support/testharness-helper.js">
</script>
<meta content="script-src 'nonce-abc' 'strict-dynamic'" http-equiv="content-security-policy"/>
<script nonce="abc">
 "use strict";

async_test(function (t) {
  assert_no_csp_event_for_url(t, "../support/import-scripts.js");
  var w = new Worker("../support/import-scripts.js");
  assert_no_event(t, w, "error");
  waitUntilEvent(w, "message").then(t.step_func_done(function (e) {
    assert_true(e.data.executed);
  }));
}, "`importScripts(...)` is allowed by 'strict-dynamic'");
</script>
